requirements.txt
streamlit==1.30.0
openai==1.12.0
pydicom==2.4.1
Pillow==10.2.0
numpy==1.26.4
plotly==5.18.0
nest_asyncio==1.6.0
streamlit_js_eval==0.1.0
